GameEvaluator interface
